Update 1 - Andrew Worthington
Added Enums Item and Challenge, to deal with features we want to add to
rooms.
Added Items Class to eventually assist in the parsing of commands that
involve items. Need to add Challenges class to do the same.
Added HashMaps to Room class to hold items and Challenges in the room,
as well as private method getLookString, to get descriptions of items
and challenges in the room, as well as public method getRoomExamination,
that returns this information along with room's description.
Added LOOK command to trigger these functions