AW Update 11/4
Added showInventory to player and Inventory command to command list.
Added carryWeight and weightLimit to player.
AddItem in player now appropriately adds to player weight, and checks
player carryWeight before allowing item to be added. This method now
returns a boolean which help determine the success of the take command.
Added the DROP command which trigger an addItem on the currentRoom after
testing player inventory and removing item if present.
Added some better formatting to LOOK command.